requirements.txt
bottle==0.13.2
cffi==1.17.1
clr_loader==0.2.7.post0
MouseInfo==0.1.3
mss==10.0.0
numpy==2.2.2
opencv-python==4.11.0.86
packaging==24.2
pillow==11.1.0
proxy_tools==0.1.0
PyAutoGUI==0.9.54
pycparser==2.22
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
pystray==0.19.5
pytesseract==0.3.13
pythonnet==3.0.5
pytweening==1.2.0
pywebview==5.4
six==1.17.0
typing_extensions==4.12.2
